refactor add mappings recreate initial migration
