+++ AppDbContext.cs
@@ -1,5 +1,10 @@
 using Microsoft.EntityFrameworkCore;
+using weather_app.Modules.Alerts.Entities;
+using weather_app.Modules.FeelsLikes.Entities;
+using weather_app.Modules.Forecasts.Entities;
+using weather_app.Modules.Temperatures.Entities;
 using weather_app.Modules.Weather.Entities;
+using weather_app.Modules.WeatherInfos.Entities;
 
 namespace weather_app
 {
@@ -16,12 +21,12 @@ namespace weather_app
         public DbSet<FeelsLike> FeelsLikes { get; set; }
         public DbSet<HourlyForecast> HourlyForecasts { get; set; }
         
-        public DbSet<MeteoAlert> MeteoAlerts { get; set; }
+        public DbSet<Alert> MeteoAlerts { get; set; }
         public DbSet<MinutelyForecast> MinutelyForecasts { get; set; }
         public DbSet<OneCallForecast> OneCallForecasts { get; set; }
         public DbSet<Temperature> Temperatures { get; set; }
         
-        public DbSet<Weather> Weather { get; set; }
+        public DbSet<WeatherInfo> Weather { get; set; }
         
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
