+++ Modules/Forecasts/Dtos/MinutelyForecastReadDto.cs
@@ -1,0 +1,13 @@
+using System.Text.Json.Serialization;
+
+namespace weather_app.Modules.Forecasts.Dtos
+{
+    public class MinutelyForecastReadDto
+    {
+        [JsonPropertyName("dt")]
+        public  long Dt { get; set; }
+        
+        [JsonPropertyName("precipitation")]
+        public double Precipitation { get; set; }
+    }
+}
