--- Modules/Weather/Dtos/TemperatureReadDto.cs
@@ -1,25 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace weather_app.Modules.Weather.Dtos
-{
-    public class TemperatureReadDto
-    {
-        [JsonPropertyName("day")]
-        public double Day { get; set; }
-        
-        [JsonPropertyName("min")]
-        public double Min { get; set; }
-        
-        [JsonPropertyName("max")]
-        public double Max { get; set; }
-        
-        [JsonPropertyName("night")]
-        public double Night { get; set; }
-        
-        [JsonPropertyName("eve")]
-        public double Eve { get; set; }
-        
-        [JsonPropertyName("morn")]
-        public double Morn { get; set; }
-    }
-}
