+++ Modules/WeatherInfos/Dtos/WeatherInfoReadDto.cs
@@ -1,0 +1,19 @@
+using System.Text.Json.Serialization;
+
+namespace weather_app.Modules.WeatherInfos.Dtos
+{
+    public class WeatherInfoReadDto
+    {
+        [JsonPropertyName("id")]
+        public int Id { get; set; }
+        
+        [JsonPropertyName("main")]
+        public string Main { get; set; }
+        
+        [JsonPropertyName("description")]
+        public string Description { get; set; }
+        
+        [JsonPropertyName("icon")]
+        public string Icon { get; set; }
+    }
+}
